Initial commit of fauxlizer parser
This parses the data generated by Fauxlizer 5000, created
at the Fantasy Analytics lab. The parser handles buggy output
that may be generated from time to time.

* rows may be truncated
* the header may be missing or malformed
* the values in a row may be in the wrong order
* there might not be any data

 The code has several functions:

 extract_data(file_name) will ensure that the required headers
 "experiment_name", "sample_id", "fauxness", "category_guess", are
 in the csv file. It will then check that
 * "experiment_name" column is not empty.
 * "sample_id" column has a positive integer ID
 * "fauxness" column is a floating point value on the range [0.0, 1.0]
   inclusive
* category_guess column is a string enum chosen from the set {"real",
  "fake", "ambiguous"}

Once the data is extracted, the data can be passed into the
generate_summary(code: str, payload: dict) function to return a summary
of the data. code is the results code. If the file is valid, it will
return a result code of "SUCCESS". Otherwise, it will return a somewhat
descriptive code to indicate the error. "payload" will contain the
resulting data if code is "SUCCESS". Otherwise, it will return the row
that the funcnction extract_data tried to parse.

fetch_row(payload, row_num, format) will return one row in the file
with the format of "format". By default, the returned data will be the
python dict of the row. Current valid formats are "JSON" and "CSV".